(FM-8178) Add Hyper-V Vagrant Support
This commit adds support to the Vagrant provisioner for
using the Hyper-V provider on windows. Right now it does
not do a smart determination of provider - if Windows,
use Hyper-V, else use VirtualBox.

It also adds three new parameters:

- `hyperv_vswitch` to specify which virtual switch the
  VM should be attached to. Note that if this is not
  specified it will attempt to use the `Default Switch`
  which is present on all Windows 10 (v1709+) machines.
- `hyperv_smb_username` and `hyperv_smb_password` which
  are required to automatically share the folder with
  the VM.

Without specifying both the switch and the SMB username
and password the task would fail as it cannot accept
any input in the context in which `bolt task run` has
to execute.

In addition to passing the parameters directly they may
also be passed as environment variables prefixed with
`LITMUS_`.

This commit updates the task, metadata, and documentation.