(FM-8274) Fix WinRM config for Windows target hosts
Prior to this commit the task used a method, `configure_ssh`, to
retrieve the connection information for a target host after vagrant
spins the machine up. It did this regardless of whether or not the
target has SSH enabled or is a Windows machine. This caused errrors
when provisioning windows targets.

Further, the WinRM config for the target used defaults which were
hardcoded and would cause nodes to never be reachable.

This commit replaces `configure_ssh()` with `configure_remoting()`
and retrieves the remoting configuration information regardless
of whether the target is using SSH or WinRM. This will also allow
us to configure WinRM on the target node if needed in the future.

NOTE: This check relies on the `platform_uses_ssh()` method which
may not be perfectly reliable; for example, a windows machine may
have OpenSSH enabled and configured.

This commit further modifies the inventory for WinRM nodes to pass
the correct username, password, and node name.

NOTE: This keeps the pre-configured default which leaves WinRM
running over HTTP. It may be worth improving this in the future.