(FM-8260) Disable vagrant synced folders by default
Prior to this commit the default behavior of vagrant machines
provisioned with the vagrant task used the synced_folder
functionality. This functionality is not used by Litmus but
might be used in other cases.

On Windows, when using the Hyper-V provider for Vagrant, it
is required to also specify a username and password for the
synced_folder. This causes provisioning runs on windows to
often fail by default if the user does not specify credentials.

This commit will cause the default behavior on Windows to
fail less while still enabling workflows that require a
synced folder.