(maint) Use a default for inventory location
Previously the provision tasks stated that the inventory parameter was optional
however if you don't pass this parameter it raises Runtime Errors as you can't
File.join on nil. This commit adds a helper to sanitise the inventory location
and return the current working directory if none is specified.

This commit also cleans up the task_helper to only require libraries it requires
instead of pulling in entire libraries (e.g. litmus) even if they're never used.