Podman mounts /tmp as tmpfs with noexec when it detects systemd inside
Podman implements a few tricks to make it easier to run systemd
container out of the box.  One of them is to mount /tmp and /run as
tmpfs with a set of default options including noexec.

However bolt runs tasks inside the container by copying in shell
scripts and executing them from the /tmp directory.