(IAC-822) Improve robustness of ABS API retry loop
Resolves an issue with the ABS provisioner where it would rapidly
retry requests to the ABS API in a quick loop, then fail.

We make a request for our VMs and exepect to receive a HTTP 202 by
the API to indicate it is servicing our request. We then go in to
a retry loop for 10 minutes, sleeping 5 secs and doubling each loop
until we hit 60 seconds. We will then break out of the loop when
either of these conditions are met:

- HTTP 200: Indicating successful provisioning
- HTTP 404: Indicating no further queries should be attempted

We will raise a unique alert for each failure condition

(IAC-822) Raise error on a HTTP 404 resp

(IAC-822) Fix inverted logic for status code check

(IAC-822) Limit sleep time to 60, not 65 secs