Make use of ABS Priority Queuing
The ABS API has the ability to set the [priority level](https://github.com/puppetlabs/always-be-scheduling#priority-queuing) of a job. When running in the context of a CI run, we will keep the priority level at 2, however, when running from a manually triggered test run, we'll request priority 1.

Fix Rubocop violation