(MAINT) Update Docker provisioner image name gsub with '.'
When attempting to provision an image from a repository with a `/`
in the name (e.g. registry.suse.com), the provisioning will fail
because docker does not support `.`s in the image name.

This change updates the `gsub` call in the image name construction
to also substitute out `.`.

We should also be escaping the `/` character too - fixes that too