(maint) Add a spec case for the provision::abs task
The abs.rb is refactored into a class that is executed only if the task
file is the executing program. This allows the spec to load the task and
manipulate the class while mocking externals. The webmock gem is used to
provide HTTP::Net mocking.

The spec is not exhaustive, but hopefully covers enough to validate that
the refactor didn't break the task, and that the bug I added in #182 is
patched.

It could be expanded, and could serve as a template for adding specs
for other tasks in the module.

It does look to expose an edge case bug of calling the provision task
with an inventory directory specified that contains a pre-existing
inventory file.