(CAT-372) - Add var support for vagrant provisioner
To use multi-node provisioning locally with vagrant we need to be able to
assign arbitrary tags to the nodes we deploy.
Other provisioners do this with a parameter called vars.

This commit introduces support for this, and essentially copies what is
already present in the other bolt tasks in this module.