!feature
Add puppetcore switch to Gemfile.

This ensures that when the PUPPET_FORGE_TOKEN is set, then the puppetcore gems are used. If this variable is not set, the bundle install will still work as normal with the public puppet and facter gems. NOTE, however, that the corresponding tests will only pass if the puppetcore gems are present on CI and used.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>