New re-usable module_ci.yml
One of the cat-github-actions re-usable workflows has been updated to switch on the puppetcore.  For reference see <https://github.com/puppetlabs/cat-github-actions/pull/129>.  This means a minor admustment is required on any modules that also need to run against puppetcore.  One, `ruby_version` must be passed in and be >= 3.1; and two, the `PUPPET_FORGE_TOKEN` must be a secret on the repository.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>